docker/v1.0/requirements.txt
numpy>=1.11.1
scipy>=0.17.0
setuptools==27.2.0
jupyter>=1.0.0
opencv-python==3.3.0.9
scikit-image>=0.12.3
scikit-learn==0.18.1
six>=1.5.2
future==0.16.0
protobuf==2.6.1
Pillow==4.2.1
h5py==2.7.0
pyyaml
progressbar
easydict
pandas
ete2
nltk
mpi4py
